backend/requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.2
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.0
pydantic==2.5.0
httpx==0.25.2
pillow==10.1.0
google-genai
opencv-python